agentic/ag_openai/pyproject.toml
[project]
name = "ag-openai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mlflow>=3.9.0",
    "openai-agents>=0.9.0",
    "python-dotenv>=1.2.1",
]
